# public/openapi.yaml
openapi: 3.1.0
info:
  title: TrustLayer API
  description: |
    **TrustLayer** is the AI Agent Action Firewall for Enterprise Systems.

    Every consequential agent action is evaluated against identity, policy, moral alignment,
    and risk rules before execution. Ed25519-signed audit certificates provide tamper-evident
    provenance for every decision.

    ## Authentication
    All API requests require a Bearer token with prefix `tl_`:
    ```
    Authorization: Bearer tl_your_api_key_here
    ```
    Generate API keys in **Account Settings → API Keys**.

    ## Tenant Isolation
    All requests require `X-Tenant-Id` header (or use the session tenant for dashboard calls).

    ## Rate Limits
    - Starter: 1,000 req/min
    - Pro: 10,000 req/min
    - Enterprise: unlimited
  version: 2.0.0
  contact:
    name: TrustLayer Support
    url: https://trustlayer.com
servers:
  - url: https://api.trustlayer.com/api/v1
    description: Production API
  - url: http://localhost:3000/api/v1
    description: Local Development
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: tl_prod_...
...
paths:
...
  /agents/{id}/rotate:
    post:
      summary: Rotate Agent Token
      description: Revokes the current JWT and issues a new one.
      tags: [Agents]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: New token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expires_at:
                    type: string
                    format: date-time
...
  /policies/packs:
    get:
      summary: List Policy Packs
      description: Returns all available regulatory policy packs with tenant installation status.
      tags: [PolicyPacks]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of policy packs
          content:
            application/json:
              schema:
                type: object
                properties:
                  packs:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        name:
                          type: string
                          example: "HIPAA Compliance Pack"
                        description:
                          type: string
                        category:
                          type: string
                          enum: [healthcare, financial, privacy, payment, security]
                        tier_required:
                          type: string
                          enum: [starter, pro, enterprise]
                        policies:
                          type: array
                          items:
                            type: object
                        installed:
                          type: boolean
                        installed_at:
                          type: string
                          format: date-time
                          nullable: true
    post:
      summary: Install Policy Pack
      description: Installs a regulatory policy pack for the tenant.
      tags: [PolicyPacks]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pack_id]
              properties:
                pack_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Pack installed
        '404':
          description: Pack not found
  /siem/export:
    get:
      summary: Pull SIEM Events
      description: |
        Pull-based SIEM export. Returns recent audit events in JSON (NDJSON) or CEF format.
        Compatible with Splunk, QRadar, Elastic SIEM, and Datadog.
      tags: [SIEM]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: format
          in: query
          schema:
            type: string
            enum: [json, cef]
            default: json
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Return events after this timestamp (default last hour)
        - name: limit
          in: query
          schema:
            type: integer
            default: 1000
            maximum: 5000
        - name: event_types
          in: query
          schema:
            type: string
          description: Comma-separated event type filter
      responses:
        '200':
          description: Audit events in requested format
          content:
            application/x-ndjson:
              schema:
                type: string
            text/plain:
              schema:
                type: string
    post:
      summary: Push SIEM Events
      description: Triggers an immediate push to all configured SIEM endpoints.
      tags: [SIEM]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Push result
          content:
            application/json:
              schema:
                type: object
                properties:
                  events_exported:
                    type: integer
                  endpoints_notified:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: string
...
  /compliance/reports:
    get:
      summary: Generate Compliance Report
      description: |
        Generates a point-in-time compliance evidence report. Covers audit trail,
        HITL oversight log, policy violations, signature coverage, and installed policy packs.

        Supported frameworks: SOC 2 Type II, HIPAA, GDPR, PCI-DSS, General.
      tags: [Compliance]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
        - name: framework
          in: query
          schema:
            type: string
            enum: [general, soc2, hipaa, gdpr, pci-dss]
            default: general
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: format
          in: query
          schema:
            type: string
            enum: [json, summary]
            default: json
      responses:
        '200':
          description: Compliance report
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  generated_at:
                    type: string
                    format: date-time
                  framework:
                    type: string
                  period:
                    type: object
                    properties:
                      from:
                        type: string
                      to:
                        type: string
                  summary:
                    type: object
  /insurance/analytics:
    get:
      summary: Risk Analytics
      description: Returns 30/60/90-day risk score trends and per-agent summaries.
      tags: [Risk]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: X-Tenant-Id
          in: header
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Risk analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  overall:
                    type: object
                    properties:
                      avg_risk_score:
                        type: number
                      max_risk_score:
                        type: number
                      total_agents:
                        type: integer
                      high_risk_agents:
                        type: integer
                      risk_trend:
                        type: string
                        enum: [improving, stable, worsening]
                  trend_30d:
                    type: array
                    items:
                      type: object
                  agent_summaries:
                    type: array
                    items:
                      type: object
...